requirements.txt
opencv-python>=4.8.1.78
numpy>=1.23.5
moviepy>=1.0.3
pillow>=10.1.0